+++ NEWS.md
@@ -5,6 +5,7 @@ New README
 - Fixed: switched from `usethis:::project_data()` to
 `usethis:::ppackage_data()`. Still not a reliable base, though, as this is still
 an internal function => TODO: get rid of this dependency
+- Fixed: use `renv2docker` instead of local package when generating templates
 
 -------------------------------------------------------------------------------